Allow passing custom request options
On the corporate environment where I am using brickster, some of the requests fail with the error message:

> Error in the HTTP2 framing layer

I have not been able to create a reproducible example of the issue. The same request sometimes triggers an error sometimes it does not.

My first suspicions went to our SSL proxy, but I requested to temporarily bypass it and I was able to reproduce the error even with the proxy disabled. I can't yet fully rule out some router/firewall somewhere that messes up some of the requests, but I'll test.

I can rule out brickster/databricks because I have triggered a similar error with a "git clone" execution.

I have found that both brickster and git use curl/libcurl as the underlying http2 client. So I can't rule out a bug in curl.

Finding the cause of the error and fixing it may be hard for me.

At this point I have a workaround for git, which is to use http 1.1. With that workaround I do not have any issue.

I would like to be able to apply the same strategy to brickster, hence this commit.

With this commit I can use:

options("brickster.request.options" = list(
  http_version = curl::curl_symbols("CURL_HTTP_VERSION_1_1")$value)
)

Long term, the best approach is to fix this http2 issue, but it can take months for me to understand the root cause and contribute a fix.

The CRAN curl package maintainer already [pointed out](https://github.com/jeroen/curl/issues/424) that setting options across all curl requests through a global variable was not an option for this use case, so my best bet right now is to implement a fix at the brickster level.